feat: Refactoring and tidy up
