+++ src/Uno.Extensions-packageonly.slnf
@@ -1,0 +1,59 @@
+{
+  "solution": {
+    "path": "Uno.Extensions.sln",
+    "projects": [
+      "Uno.Extensions.Authentication.MSAL\\Uno.Extensions.Authentication.MSAL.UI.Wasm.csproj",
+      "Uno.Extensions.Authentication.MSAL\\Uno.Extensions.Authentication.MSAL.UI.csproj",
+      "Uno.Extensions.Authentication.MSAL\\Uno.Extensions.Authentication.MSAL.WinUI.Wasm.csproj",
+      "Uno.Extensions.Authentication.MSAL\\Uno.Extensions.Authentication.MSAL.WinUI.csproj",
+      "Uno.Extensions.Authentication.Oidc\\Uno.Extensions.Authentication.Oidc.UI.csproj",
+      "Uno.Extensions.Authentication.Oidc\\Uno.Extensions.Authentication.Oidc.WinUI.csproj",
+      "Uno.Extensions.Authentication.UI\\Uno.Extensions.Authentication.UI.csproj",
+      "Uno.Extensions.Authentication.UI\\Uno.Extensions.Authentication.WinUI.csproj",
+      "Uno.Extensions.Authentication\\Uno.Extensions.Authentication.csproj",
+      "Uno.Extensions.Configuration\\Uno.Extensions.Configuration.csproj",
+      "Uno.Extensions.Core.Generators\\Uno.Extensions.Core.Generators.csproj",
+      "Uno.Extensions.Core\\Uno.Extensions.Core.csproj",
+      "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.UWP.Skia.csproj",
+      "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.UWP.Wasm.csproj",
+      "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.UWP.csproj",
+      "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.WinUI.Skia.csproj",
+      "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.WinUI.Wasm.csproj",
+      "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.WinUI.csproj",
+      "Uno.Extensions.Hosting\\Uno.Extensions.Hosting.csproj",
+      "Uno.Extensions.Http.Refit\\Uno.Extensions.Http.Refit.csproj",
+      "Uno.Extensions.Http.Tests\\Uno.Extensions.Http.Tests.csproj",
+      "Uno.Extensions.Http\\Uno.Extensions.Http.csproj",
+      "Uno.Extensions.Localization.UI\\Uno.Extensions.Localization.UI.csproj",
+      "Uno.Extensions.Localization.UI\\Uno.Extensions.Localization.WinUI.csproj",
+      "Uno.Extensions.Localization\\Uno.Extensions.Localization.csproj",
+      "Uno.Extensions.Logging.Serilog\\Uno.Extensions.Logging.Serilog.csproj",
+      "Uno.Extensions.Logging\\Uno.Extensions.Logging.UWP.Skia.csproj",
+      "Uno.Extensions.Logging\\Uno.Extensions.Logging.UWP.Wasm.csproj",
+      "Uno.Extensions.Logging\\Uno.Extensions.Logging.UWP.csproj",
+      "Uno.Extensions.Logging\\Uno.Extensions.Logging.WinUI.Skia.csproj",
+      "Uno.Extensions.Logging\\Uno.Extensions.Logging.WinUI.Wasm.csproj",
+      "Uno.Extensions.Logging\\Uno.Extensions.Logging.WinUI.csproj",
+      "Uno.Extensions.Navigation.Toolkit\\Uno.Extensions.Navigation.Toolkit.UI.csproj",
+      "Uno.Extensions.Navigation.Toolkit\\Uno.Extensions.Navigation.Toolkit.WinUI.csproj",
+      "Uno.Extensions.Navigation.UI\\Uno.Extensions.Navigation.UI.csproj",
+      "Uno.Extensions.Navigation.UI\\Uno.Extensions.Navigation.WinUI.csproj",
+      "Uno.Extensions.Navigation.WinUI.Markup\\Uno.Extensions.Navigation.WinUI.Markup.csproj",
+      "Uno.Extensions.Navigation\\Uno.Extensions.Navigation.csproj",
+      "Uno.Extensions.Reactive.Generator\\Uno.Extensions.Reactive.Generator.csproj",
+      "Uno.Extensions.Reactive.Messaging\\Uno.Extensions.Reactive.Messaging.csproj",
+      "Uno.Extensions.Reactive.UI\\Uno.Extensions.Reactive.UI.WinUI.csproj",
+      "Uno.Extensions.Reactive.UI\\Uno.Extensions.Reactive.UI.csproj",
+      "Uno.Extensions.Reactive.WinUI.Markup\\Uno.Extensions.Reactive.WinUI.Markup.csproj",
+      "Uno.Extensions.Reactive\\Uno.Extensions.Reactive.csproj",
+      "Uno.Extensions.Serialization.Http\\Uno.Extensions.Serialization.Http.csproj",
+      "Uno.Extensions.Serialization.Refit\\Uno.Extensions.Serialization.Refit.csproj",
+      "Uno.Extensions.Serialization\\Uno.Extensions.Serialization.csproj",
+      "Uno.Extensions.Storage.UI\\Uno.Extensions.Storage.UI.csproj",
+      "Uno.Extensions.Storage.UI\\Uno.Extensions.Storage.WinUI.csproj",
+      "Uno.Extensions.Storage\\Uno.Extensions.Storage.csproj",
+      "Uno.Extensions.Templates\\Uno.Extensions.Templates.csproj",
+      "Uno.Extensions\\Uno.Extensions.csproj"
+    ]
+  }
+}
